Add W-9 upload and admin review to contractor applications
Implements a secure W-9 PDF upload step in the contractor application process, storing the file in Firebase Storage and saving the URL in Firestore. Updates the admin applications review UI to display a W-9 status badge and a link to view the uploaded W-9. Updates types and review steps to support the new field. Also adds the latest IRS Form W-9 PDF to the repository.